+++ src/pages/SignIn/SignIn.jsx
@@ -3,11 +3,7 @@ import { styled } from "styled-components";
 import Signinform from "~components/auth/Signin/Signinform";
 
 function SignIn() {
-  return (
-    <div>
-      <Signinform />
-    </div>
-  );
+  return <Signinform />;
 }
 
 export default SignIn;
